Cleaned up subvoxel files, moved subvoxel-specific buffers to own section
